Fixed Authentication and Gitignore
For some reason the gitignore wasn't working as needed. Fixed it.

Before there were some errors authenticating. First of all there was no
error handler defined. Got that done. Second we added a pipeline to
incorporate that error handler ( @ config ). Third we fixed the
references you know, because guardian updated and we needed to be
referring to quoteboard.guardian instead of just guardian. We also
updated the routes to incorporate these changes. Ca-ching