Updated to work with Docker-Compose
Created a new docker-compose file and updated commands accordingly. This
should be adequate for quick and easy development. It utilizes my own
kritr/phoenix image that comes with buildtools so Argon2 can compile. I
also made that its own image just so it would be available for download
rather than needing to be built.
The mix file has also been updated with an additional task to get deps
and serve the repository (the command run when docker compose is
launched).
In addition, the tmp folder has been added to the ignore file.